requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
boto3==1.34.108
botocore==1.34.108
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
Django==5.0.6
django-browser-reload==1.12.1
django-storages==1.14.3
django-tailwind==3.8.0
dotenv-python==0.0.1
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pillow==10.3.0
Pygments==2.18.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
rich==13.7.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
